+++ common/item_data.h
@@ -541,8 +541,8 @@ namespace EQ
 		char	ScrollName[65] {};
 		//BardName
 
-		bool IsEquipable(uint16 Race, uint16 Class) const;
-		bool IsClassEquipable(uint16 Class) const;
+		bool IsEquipable(uint16 Race, uint32 Class) const;
+		bool IsClassEquipable(uint32 Class) const;
 		bool IsRaceEquipable(uint16 Race) const;
 		bool IsClassCommon() const;
 		bool IsClassBag() const;
