+++ common/item_instance.h
@@ -98,8 +98,8 @@ namespace EQ
 		bool IsCharged() const;
 
 		// Can item be equipped by/at?
-		bool IsEquipable(uint16 race, uint16 class_) const;
-		bool IsClassEquipable(uint16 class_) const;
+		bool IsEquipable(uint16 race, uint32 class_) const;
+		bool IsClassEquipable(uint32 class_) const;
 		bool IsRaceEquipable(uint16 race) const;
 		bool IsEquipable(int16 slot_id) const;
